Move Metabox en Sidebar slots, filters, and portals into their own concepts.
