+++ js/src/components/fills/SidebarFill.js
@@ -1,39 +1,38 @@
 /* External dependencies */
 import { Fill } from "@wordpress/components";
 import { Fragment } from "@wordpress/element";
-import { __ } from "@wordpress/i18n";
 import PropTypes from "prop-types";
 
 /* Internal dependencies */
-import CollapsibleCornerstone from "../containers/CollapsibleCornerstone";
-import SnippetEditor from "../containers/SnippetEditor";
-import Warning from "../containers/Warning";
-import KeywordInput from "./contentAnalysis/KeywordInput";
-import ReadabilityAnalysis from "./contentAnalysis/ReadabilityAnalysis";
-import SeoAnalysis from "./contentAnalysis/SeoAnalysis";
-import Collapsible from "./SidebarCollapsible";
-import SidebarItem from "./SidebarItem";
-import TopLevelProviders from "./TopLevelProviders";
-import Social from "./SocialMetadata";
+import CollapsibleCornerstone from "../../containers/CollapsibleCornerstone";
+import Warning from "../../containers/Warning";
+import KeywordInput from "../contentAnalysis/KeywordInput";
+import ReadabilityAnalysis from "../contentAnalysis/ReadabilityAnalysis";
+import SeoAnalysis from "../contentAnalysis/SeoAnalysis";
+import SidebarItem from "../SidebarItem";
+import SnippetPreviewModal from "../SnippetPreviewModal";
+import TopLevelProviders from "../TopLevelProviders";
 
 /**
- * Creates the Metabox component.
+ * Creates the SidebarFill component.
  *
  * @param {Object} settings The feature toggles.
  * @param {Object} store    The Redux store.
  * @param {Object} theme    The theme to use.
  *
- * @returns {ReactElement} The Metabox component.
+ * @returns {ReactElement} The Sidebar component.
+ *
+ * @constructor
  */
-export default function Metabox( { settings, store, theme } ) {
+export default function SidebarFill( { settings, store, theme } ) {
 	return (
 		<Fragment>
-			<Fill name="YoastMetabox">
+			<Fill name="YoastSidebar">
 				<SidebarItem renderPriority={ 1 }>
 					<TopLevelProviders
 						store={ store }
 						theme={ theme }
-						location={ "metabox" }
+						location={ "sidebar" }
 					>
 						<Warning />
 					</TopLevelProviders>
@@ -42,30 +41,25 @@ export default function Metabox( { settings, store, theme } ) {
 					<TopLevelProviders
 						store={ store }
 						theme={ theme }
-						location={ "metabox" }
+						location={ "sidebar" }
 					>
 						<KeywordInput />
 					</TopLevelProviders>
 				</SidebarItem> }
-				<SidebarItem renderPriority={ 9 }>
+				{ <SidebarItem renderPriority={ 9 }>
 					<TopLevelProviders
 						store={ store }
 						theme={ theme }
-						location={ "metabox" }
+						location={ "sidebar" }
 					>
-						<Collapsible
-							id={ "yoast-snippet-editor-metabox" }
-							title={ __( "Google preview", "wordpress-seo" ) } initialIsOpen={ true }
-						>
-							<SnippetEditor hasPaperStyle={ false } />
-						</Collapsible>
+						<SnippetPreviewModal />
 					</TopLevelProviders>
-				</SidebarItem>
+				</SidebarItem> }
 				{ settings.isContentAnalysisActive && <SidebarItem renderPriority={ 10 }>
 					<TopLevelProviders
 						store={ store }
 						theme={ theme }
-						location={ "metabox" }
+						location={ "sidebar" }
 					>
 						<ReadabilityAnalysis />
 					</TopLevelProviders>
@@ -74,7 +68,7 @@ export default function Metabox( { settings, store, theme } ) {
 					<TopLevelProviders
 						store={ store }
 						theme={ theme }
-						location={ "metabox" }
+						location={ "sidebar" }
 					>
 						<SeoAnalysis
 							shouldUpsell={ settings.shouldUpsell }
@@ -86,24 +80,18 @@ export default function Metabox( { settings, store, theme } ) {
 					<TopLevelProviders
 						store={ store }
 						theme={ theme }
-						location={ "metabox" }
+						location={ "sidebar" }
 					>
 						<CollapsibleCornerstone />
 					</TopLevelProviders>
-				</SidebarItem> }
-				<TopLevelProviders
-					store={ store }
-					theme={ theme }
-					location={ "metabox" }
-				>
-					<Social />
-				</TopLevelProviders>
+				</SidebarItem>
+				}
 			</Fill>
 		</Fragment>
 	);
 }
 
-Metabox.propTypes = {
+SidebarFill.propTypes = {
 	settings: PropTypes.object.isRequired,
 	store: PropTypes.object.isRequired,
 	theme: PropTypes.object.isRequired,
